// briefly-orchestration/src/mastra/types/episodeStyleClassification.ts
export enum EpisodeStyle {
  EVERGREEN_NARRATIVE = "EVERGREEN_NARRATIVE",
  EXPLAINER_DEEP_DIVE = "EXPLAINER_DEEP_DIVE",
  BREAKING_RECENT_DEVELOPMENTS = "BREAKING_RECENT_DEVELOPMENTS",
  HEADLINES_DIGEST = "HEADLINES_DIGEST",
  LOCAL_TIME_BOUND_EVENTS = "LOCAL_TIME_BOUND_EVENTS",
  OPINION_SPECTRUM_PERSPECTIVE_BALANCE = "OPINION_SPECTRUM_PERSPECTIVE_BALANCE",
  TREND_PATTERN_ANALYSIS = "TREND_PATTERN_ANALYSIS",
  PRACTICAL_GUIDANCE_HOW_TO = "PRACTICAL_GUIDANCE_HOW_TO",
  CREATIVE_EXPLORATORY = "CREATIVE_EXPLORATORY",
}